移除 JsonPropertiesUtils,将相应的方法移至IOUtils下,修复JSONFeidDemo2 test_0,原因是 fastjson.properties 内 fastjson.compatibleWithFieldName=true 导致的
