+++ src/main/java/com/alibaba/fastjson/util/IOUtils.java
@@ -16,6 +16,7 @@
 package com.alibaba.fastjson.util;
 
 import java.io.Closeable;
+import java.io.InputStream;
 import java.nio.ByteBuffer;
 import java.nio.CharBuffer;
 import java.nio.charset.CharacterCodingException;
@@ -23,7 +24,10 @@ import java.nio.charset.Charset;
 import java.nio.charset.CharsetDecoder;
 import java.nio.charset.CoderResult;
 import java.nio.charset.MalformedInputException;
+import java.security.AccessController;
+import java.security.PrivilegedAction;
 import java.util.Arrays;
+import java.util.Properties;
 
 import com.alibaba.fastjson.JSONException;
 
@@ -31,6 +35,14 @@ import com.alibaba.fastjson.JSONException;
  * @author wenshao[szujobs@hotmail.com]
  */
 public class IOUtils {
+    
+    public  final  static String FASTJSON_PROPERTIES  ="fastjson.properties";
+    
+    public final static String FASTJSON_COMPATIBLEWITHJAVABEAN="fastjson.compatibleWithJavaBean";
+    
+    public final static String FASTJSON_COMPATIBLEWITHFIELDNAME="fastjson.compatibleWithFieldName";
+    
+    public final static Properties DEFAULT_PROPERTIES =new Properties();    
 
     public final static Charset   UTF8                 = Charset.forName("UTF-8");
     
@@ -62,6 +74,55 @@ public class IOUtils {
                 identifierFlags[c] = true;
             } else if (c >= '0' && c <= '9') {
                 identifierFlags[c] = true;
+            }
+        }
+    }
+    
+    static {
+        try {
+            new PropertiesInitializer().autoConfig();
+        } catch (Exception e) {
+            //skip
+        }
+    }
+    
+    
+    static class PropertiesInitializer{
+        public void autoConfig(){
+            loadPropertiesFromFile();
+            TypeUtils.compatibleWithJavaBean ="true".equals(getStringProperty(FASTJSON_COMPATIBLEWITHJAVABEAN)) ;
+            TypeUtils.compatibleWithFieldName ="true".equals(getStringProperty(FASTJSON_COMPATIBLEWITHFIELDNAME)) ;
+        }
+    }
+    
+    public static String getStringProperty(String name) {
+        String prop = null;
+        try {
+            prop = System.getProperty(name);
+        } catch (SecurityException e) {
+            ; // skip
+        }
+        return (prop == null) ? DEFAULT_PROPERTIES.getProperty(name) : prop;
+    }
+    
+    public static void loadPropertiesFromFile(){
+        InputStream imputStream = AccessController.doPrivileged(new PrivilegedAction<InputStream>() {
+            public InputStream run() {
+                ClassLoader cl = Thread.currentThread().getContextClassLoader();
+                if (cl != null) {
+                    return cl.getResourceAsStream(FASTJSON_PROPERTIES);
+                } else {
+                    return ClassLoader.getSystemResourceAsStream(FASTJSON_PROPERTIES);
+                }
+            }
+        });
+        
+        if (null != imputStream) {
+            try {
+                DEFAULT_PROPERTIES.load(imputStream);
+                imputStream.close();
+            } catch (java.io.IOException e) {
+                // skip
             }
         }
     }
