+++ src/test/resources/fastjson.properties
@@ -1,6 +1,6 @@
 #default
 fastjson.compatibleWithJavaBean=false
 #default
-fastjson.compatibleWithFieldName=true
+fastjson.compatibleWithFieldName=false
 #default
 fastjson.parser.deny=
